Python Reverse Shell
Hi, this is a reverse shell program that helps you send commands in DOS
to a Windows machine on the LAN (these were tested).
The machine that sould run the commands is the client and the sender is
the server.
There are commands like:
screenshot, cd, grab* and so on which are coded in the program for
easyer use.
I will be glad to recieve some ideas for making it better.

*I will not take any responsibility for any bad use of it.